Resolved apparently random test failure displaying as a bind exception on the embedded LDAP server, when using real service classes in integration tests
