+++ src/main/java/uk/gov/justice/digital/delius/controller/advice/GeneralControllerAdvice.java
@@ -1,0 +1,30 @@
+package uk.gov.justice.digital.delius.controller.advice;
+
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.ControllerAdvice;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.client.HttpClientErrorException;
+import org.springframework.web.client.HttpServerErrorException;
+import uk.gov.justice.digital.delius.service.NoSuchUserException;
+
+import static org.springframework.http.HttpStatus.NOT_FOUND;
+
+@ControllerAdvice
+public class GeneralControllerAdvice {
+
+   @ExceptionHandler(HttpClientErrorException.class)
+   public ResponseEntity<String> restClientError(HttpClientErrorException e) {
+       return new ResponseEntity<>(e.getMessage(), e.getStatusCode());
+   }
+
+   @ExceptionHandler(HttpServerErrorException.class)
+   public ResponseEntity<String> restServerError(HttpServerErrorException e) {
+       return new ResponseEntity<>(e.getMessage(), e.getStatusCode());
+   }
+
+   @ExceptionHandler(NoSuchUserException.class)
+   public ResponseEntity<String> noSuchUser(NoSuchUserException e) {
+       return new ResponseEntity<>(e.getMessage(), NOT_FOUND);
+   }
+}
+
